Fix: Remove BaseTestCase
