+++ tests/Integration/Domain/Model/AirportTest.php
@@ -15,10 +15,10 @@ namespace OpenCFP\Test\Integration\Domain\Model;
 
 use OpenCFP\Domain\EntityNotFoundException;
 use OpenCFP\Domain\Model\Airport;
-use OpenCFP\Test\BaseTestCase;
 use OpenCFP\Test\Helper\RefreshDatabase;
+use OpenCFP\Test\Integration\WebTestCase;
 
-final class AirportTest extends BaseTestCase
+final class AirportTest extends WebTestCase
 {
     use RefreshDatabase;
