+++ tests/Integration/Domain/Model/TalkTest.php
@@ -16,10 +16,10 @@ namespace OpenCFP\Test\Integration\Domain\Model;
 use OpenCFP\Domain\Model\Favorite;
 use OpenCFP\Domain\Model\Talk;
 use OpenCFP\Domain\Model\TalkMeta;
-use OpenCFP\Test\BaseTestCase;
 use OpenCFP\Test\Helper\RefreshDatabase;
+use OpenCFP\Test\Integration\WebTestCase;
 
-final class TalkTest extends BaseTestCase
+final class TalkTest extends WebTestCase
 {
     use RefreshDatabase;
